+++ java/src/main/java/org/rocksdb/ColumnFamilyOptionsInterface.java
@@ -456,7 +456,7 @@ public interface ColumnFamilyOptionsInterface {
    *     files and the total size of level-L files for all L.
    * @return the reference to the current option.
    */
-  Object setMaxBytesForLevelMultiplier(int multiplier);
+  Object setMaxBytesForLevelMultiplier(double multiplier);
 
   /**
    * The ratio between the total size of level-(L+1) files and the total
@@ -466,7 +466,7 @@ public interface ColumnFamilyOptionsInterface {
    * @return the ratio between the total size of level-(L+1) files and
    *     the total size of level-L files for all L.
    */
-  int maxBytesForLevelMultiplier();
+  double maxBytesForLevelMultiplier();
 
   /**
    * Maximum size of each compaction (not guarantee)