+++ java/src/test/java/org/rocksdb/ColumnFamilyOptionsTest.java
@@ -172,9 +172,9 @@ public class ColumnFamilyOptionsTest {
   @Test
   public void maxBytesForLevelMultiplier() {
     try (final ColumnFamilyOptions opt = new ColumnFamilyOptions()) {
-      final int intValue = rand.nextInt();
-      opt.setMaxBytesForLevelMultiplier(intValue);
-      assertThat(opt.maxBytesForLevelMultiplier()).isEqualTo(intValue);
+      final double doubleValue = rand.nextDouble();
+      opt.setMaxBytesForLevelMultiplier(doubleValue);
+      assertThat(opt.maxBytesForLevelMultiplier()).isEqualTo(doubleValue);
     }
   }